Add Alarm for Rabbit Queued Messages
This commit adds an alarm with configurable threshold for rabbitmq
queued messages. Support identified that in some situations messages are
published but not consumed, this alarm should detect that situation.

The default threshold is 100 queued messages, this can be tuned via user
variables if necessary for larger deployments.

Note that originally there was a request to exclude notifications from
the count of messages as they may not be consumed so could skew the
value. However a better remedy in that case would is to
disable notifications via the noop driver for each server.

Closes #633
Trello: https://trello.com/c/widYAV5F/35-add-new-monitor-rabbitmq-queue-length