Remove incorrect run_queues test
The run_queues metric is load metric from the erlang environment, it
isn't directly related to rabbitmq message queues, though it is related
to node load, so a busy broker may have a higher run_queues.

However as this is a measure of load, it is not a good value for
checking replication across rabbit cluster members. This test has been
removed.

This patch also adds a couple of checks to ensure that all nodes are
clustered succesfully.

Related: #626