Give maas rabbit user perms on all vhosts
Previously the maas user only had permisions on the vhost "/" which is
no longer used by the openstack services, as they each have their own
vhost.

No modification to the plugin is necessary, as the same api calls will
return data from all the vhosts the user has access to.

Closes #1113